feat: 파일명 변경 및 메직넘버 제거
